requirements.txt
click>=6.2
azure-mgmt-resource>=1.1.0
azure-mgmt-compute>=1.0.0
